blogging_api/requirements.txt
django
python-decouple
gunicorn
redis
django-redis
drf-yasg
django-restframework